requirements.txt
# utils
joblib==1.0.0 #0.17.0
# pylint==2.7.4
# tqdm==4.60.0
# hydra-optuna-sweeper==1.1.0
# hydra-nevergrad-sweeper==1.1.5
# hydra-ax-sweeper==1.1.5
# hydra-colorlog==1.1.0
omegaconf==2.1.0
# hydra==1.1.0
# numba==0.53.1

# signal
heartpy==1.2.7
# librosa==0.8.1
scikit-image==0.18.2
lightgbm==3.2.1
wfdb==3.4.0

# ml
MarkupSafe==2.0
mlflow==1.18.0
# scikit-multilearn==0.2.0
# imbalanced-learn==0.8.0
# xgboost==1.4.2
# pycaret==2.3.0
neurokit2==0.1.0
# jupyterlab==3.0.14
# scikit-learn==0.23.2
# pandas==1.2.4
protobuf==3.9.2
numpy==1.20.3
pandas==1.5.0

# torch
# torch==1.8.0
pytorch-lightning==1.3.5
# torchvision==0.9.0 # to use mlflow autolog pytorch-lightning models
# torchaudio==0.8.0

# data visualization
coloredlogs==15.0
# seaborn==0.11.1
# ipympl==0.7.0
# sweetviz==2.0.9
# pandas_profiling[notebook]==2.11.0
# xlsxwriter==1.4.2
torchmetrics==0.6.1
